requirements.txt
ultralytics
torch
torchvision
customtkinter
pillow
numpy
opencv-python
pandas
openpyxl
